Update ubuntu AMIs for spectre/meltdown fixes
This commit updates the default AMIs used for orchestration to use
Ubuntu 16.04 releases that include spectre and meltdown fixes.

This can be confirmed by cloning:

https://github.com/speed47/spectre-meltdown-checker

and running:

sudo ./spectre-meltdown-checker.sh

NOTE: Currently, on our usual c4.8xlarge instance type in AWS, this
checker indicates that Spectre V2 is not fixed (no retpoline and no
updated firmware).

NOTE 2: Currently, on our usual c4.8xlarge instance type in AWS, this
checker indicates the following:

  The microcode your CPU is running on is known to cause instability problems,
  such as intempestive reboots or random crashes.
  You are advised to either revert to a previous microcode version (that might not have
  the mitigations for Spectre), or upgrade to a newer one if available.

We do NOT have control over CPU microcode (as far as I know) so we have
to rely on AWS/EC2 to resolve this.